requirements.txt
streamlit>=1.26.0
transformers
faiss-cpu
torch
pandas
kaggle
pydrive
oauth2client
pillow